+++ examples/esp32/apps/memfault_demo_app/main/memfault/memfault_platform_config.h
@@ -9,7 +9,6 @@
 //! Default configuration settings can be found in "memfault/config.h"
 
 #define MEMFAULT_TASK_WATCHDOG_ENABLE 1
-#define MEMFAULT_COMPACT_LOG_ENABLE 1
 
 // Enable the sync_successful metric
 #define MEMFAULT_METRICS_SYNC_SUCCESS 1