+++ ports/zephyr/include/memfault/ports/zephyr/core.h
@@ -1,0 +1,18 @@
+#pragma once
+
+//! @file
+//!
+//! Copyright (c) Memfault, Inc.
+//! See License.txt for details
+//!
+//! Core Zephyr Memfault port module.
+
+#ifdef __cplusplus
+extern "C" {
+#endif
+
+void memfault_zephyr_collect_reset_info(void);
+
+#ifdef __cplusplus
+}
+#endif