+++ tests/src/test_memfault_self_test.cpp
@@ -29,6 +29,10 @@ void memfault_self_test_reboot_reason_test(void) {
 uint32_t memfault_self_test_reboot_reason_test_verify(void) {
   return mock().actualCall(__func__).returnUnsignedIntValue();
 }
+
+uint32_t memfault_self_test_time_test(void) {
+  return mock().actualCall(__func__).returnUnsignedIntValue();
+}
 }
 
 // Helper function for tests that return a result rather than just running
@@ -85,11 +89,16 @@ TEST(MemfaultSelfTest, Test_RebootReasonTestVerify) {
                                 kMemfaultSelfTestFlag_RebootReasonVerify);
 }
 
+TEST(MemfaultSelfTest, Test_PlatformTimeTest) {
+  prv_run_single_component_test("memfault_self_test_time_test", kMemfaultSelfTestFlag_PlatformTime);
+}
+
 TEST(MemfaultSelfTest, Test_SelfTestDefaultHappyPath) {
   mock().expectOneCall("memfault_self_test_device_info_test").andReturnValue(0);
   mock().expectOneCall("memfault_self_test_component_boot_test").andReturnValue(0);
   mock().expectOneCall("memfault_self_test_coredump_regions_test").andReturnValue(0);
   mock().expectOneCall("memfault_self_test_data_export_test");
+  mock().expectOneCall("memfault_self_test_time_test").andReturnValue(0);
 
   int result = memfault_self_test_run(kMemfaultSelfTestFlag_Default);
   LONGS_EQUAL(0, result);
@@ -100,6 +109,7 @@ TEST(MemfaultSelfTest, Test_SelfTestDefaultComponentBootFail) {
   mock().expectOneCall("memfault_self_test_component_boot_test").andReturnValue(1);
   mock().expectOneCall("memfault_self_test_coredump_regions_test").andReturnValue(0);
   mock().expectOneCall("memfault_self_test_data_export_test");
+  mock().expectOneCall("memfault_self_test_time_test").andReturnValue(0);
 
   int result = memfault_self_test_run(kMemfaultSelfTestFlag_Default);
   LONGS_EQUAL(1, result);